requirements.txt
camel-ai
camel-ai[all]
camel-ai[huggingface-agent]
camel-ai[tools]
python-pptx
streamlit
plotly
google-generativeai
deepseek
openai
huggingface-hub
python-dotenv
kaleido






